make narrowing expansion parameter configurable
this grows the grid a little after it gets narrowed down.
That makes targets that happen to be on lines between
grid fields easier to hit.

Also allows the grid to have a minimum size, which
lets you nudge it all around the screen (very slowly) when
it's at its smallest size.